:recycle: Rename types Module To Avoid Shadowing
Avoid shadowing the standard library types and typing modules.
