+++ wsic/__init__.py
@@ -4,14 +4,14 @@ __author__ = """John Pocock"""
 __email__ = "j.c.pocock@warwick.ac.uk"
 __version__ = "0.6.1"
 
-from . import codecs, magic, metadata, multiproc, readers, types, writers
+from . import codecs, magic, metadata, multiproc, readers, typedefs, writers
 
 __all__ = [
     "codecs",
     "magic",
-    "readers",
-    "types",
-    "writers",
     "metadata",
     "multiproc",
+    "readers",
+    "typedefs",
+    "writers",
 ]
